requirements_1llll.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
Django==5.1.5
django-celery-beat==2.7.0
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-timezone-field==7.1
Faker==36.1.1
fonttools==4.56.0
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
razorpay==1.4.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
shapely==2.0.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
